+++ example/lib/main.dart
@@ -19,14 +19,14 @@ class App extends StatelessWidget {
   late final _router = GoRouter(routes: _routesBuilder, error: _errorBuilder);
   List<GoRoute> _routesBuilder(BuildContext context, String location) => [
         GoRoute(
-          pattern: '/',
+          path: '/',
           builder: (context, state) => MaterialPage<Page1Page>(
             key: state.pageKey,
             child: const Page1Page(),
           ),
         ),
         GoRoute(
-          pattern: '/page2',
+          path: '/page2',
           builder: (context, state) => MaterialPage<Page2Page>(
             key: state.pageKey,
             child: const Page2Page(),