+++ example/lib/query_params.dart
@@ -27,19 +27,26 @@ class App extends StatelessWidget {
   late final _router = GoRouter(
     routes: _routeBuilder,
     error: _errorBuilder,
-    guard: Guard(loginInfo),
+
+    // the guard checks if the user is logged in via the GoRouterLoggedIn mixin,
+    // passing in the original location as a query parameter
+    guard: GoRouterLoginGuard(
+      loginInfo,
+      loginPath: '/login',
+      fromParam: 'from',
+    ),
   );
 
   List<GoRoute> _routeBuilder(BuildContext context, String location) => [
         GoRoute(
-          pattern: '/',
+          path: '/',
           builder: (context, state) => MaterialPage<HomePage>(
             key: state.pageKey,
             child: HomePage(families: Families.data),
           ),
           routes: [
             GoRoute(
-              pattern: 'family/:fid',
+              path: 'family/:fid',
               builder: (context, state) {
                 final family = Families.family(state.params['fid']!);
                 return MaterialPage<FamilyPage>(
@@ -49,7 +56,7 @@ class App extends StatelessWidget {
               },
               routes: [
                 GoRoute(
-                  pattern: 'person/:pid',
+                  path: 'person/:pid',
                   builder: (context, state) {
                     final family = Families.family(state.params['fid']!);
                     final person = family.person(state.params['pid']!);
@@ -64,7 +71,7 @@ class App extends StatelessWidget {
           ],
         ),
         GoRoute(
-          pattern: '/login',
+          path: '/login',
           builder: (context, state) => MaterialPage<LoginPage>(
             key: state.pageKey,
             // pass the deep link to the LoginPage (if there is one)
@@ -78,30 +85,4 @@ class App extends StatelessWidget {
         key: state.pageKey,
         child: ErrorPage(state.error),
       );
-}
-
-class Guard extends GoRouterGuard {
-  // passing loginInfo to the base class will cause a change to trigger routing
-  Guard(LoginInfo loginInfo) : super(loginInfo);
-
-  LoginInfo get loginInfo => super.listenable! as LoginInfo;
-
-  // redirect based on app and routing state
-  @override
-  String? redirect(String location) {
-    // watching LoginInfo will cause a change in LoginInfo to trigger routing
-    final loggedIn = loginInfo.loggedIn;
-    final loc = Uri.parse(location).path;
-    final goingToLogin = loc == '/login';
-
-    // the user is not logged in and not headed to /login, they need to login
-    // also, keep track of the deep link that was used to get here
-    if (!loggedIn && !goingToLogin) return '/login?from=$loc';
-
-    // the user is logged in and headed to /login, no need to login again
-    if (loggedIn && goingToLogin) return '/';
-
-    // no need to redirect at all
-    return null;
-  }
 }