+++ example/lib/redirection.dart
@@ -27,19 +27,21 @@ class App extends StatelessWidget {
   late final _router = GoRouter(
     routes: _routesBuilder,
     error: _errorBuilder,
-    guard: Guard(loginInfo), // the guard checks if the user is logged in
+
+    // the guard checks if the user is logged in via the GoRouterLoggedIn mixin
+    guard: GoRouterLoginGuard(loginInfo, loginPath: '/login'),
   );
 
   List<GoRoute> _routesBuilder(BuildContext context, String location) => [
         GoRoute(
-          pattern: '/',
+          path: '/',
           builder: (context, state) => MaterialPage<HomePage>(
             key: state.pageKey,
             child: HomePage(families: Families.data),
           ),
           routes: [
             GoRoute(
-              pattern: 'family/:fid',
+              path: 'family/:fid',
               builder: (context, state) {
                 final family = Families.family(state.params['fid']!);
                 return MaterialPage<FamilyPage>(
@@ -49,7 +51,7 @@ class App extends StatelessWidget {
               },
               routes: [
                 GoRoute(
-                  pattern: 'person/:pid',
+                  path: 'person/:pid',
                   builder: (context, state) {
                     final family = Families.family(state.params['fid']!);
                     final person = family.person(state.params['pid']!);
@@ -64,7 +66,7 @@ class App extends StatelessWidget {
           ],
         ),
         GoRoute(
-          pattern: '/login',
+          path: '/login',
           builder: (context, state) => MaterialPage<LoginPage>(
             key: state.pageKey,
             child: const LoginPage(),
@@ -77,27 +79,4 @@ class App extends StatelessWidget {
         key: state.pageKey,
         child: ErrorPage(state.error),
       );
-}
-
-class Guard extends GoRouterGuard {
-  // passing loginInfo to the base class will cause a change to trigger routing
-  Guard(LoginInfo loginInfo) : super(loginInfo);
-
-  LoginInfo get loginInfo => super.listenable! as LoginInfo;
-
-  // redirect based on app and routing state
-  @override
-  String? redirect(String location) {
-    final loggedIn = loginInfo.loggedIn;
-    final goingToLogin = location == '/login';
-
-    // the user is not logged in and not headed to /login, they need to login
-    if (!loggedIn && !goingToLogin) return '/login';
-
-    // the user is logged in and headed to /login, no need to login again
-    if (loggedIn && goingToLogin) return '/';
-
-    // no need to redirect at all
-    return null;
-  }
 }